requirements.txt
# API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0

# Worker
celery[redis]==5.3.4
redis==5.0.1

# Storage
minio==7.2.0

# Computer Vision
opencv-python-headless==4.8.1.78
ultralytics==8.0.200
Pillow==10.1.0
numpy==1.26.2

# Video processing
ffmpeg-python==0.2.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
